agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.55.1",
    "phidata>=2.5.33",
    "fastapi-mail>=1.4.1",
    "fastapi[standard]>=0.112.2",
    "logfire[fastapi,sqlalchemy]>=0.51.0",
    "sqlalchemy>=2.0.32",
    "supabase>=2.7.3",
    "python-dotenv>=1.0.1",
    "duckduckgo-search>=6.3.7",
    "pydantic>=2.10.2",
    "arxiv>=2.1.3",
    "pypdf>=5.1.0",
    "ipykernel>=6.29.5",
    "pgvector>=0.3.6",
    "langchain>=0.3.8",
    "langchain-community>=0.3.8",
    "langchain-openai>=0.2.10",
    "unstructured>=0.16.8",
    "crewai[tools]>=0.83.0",
    "markdown>=3.7",
    "aiosmtplib>=3.0.2",
    "pydantic-settings>=2.6.1",
]
